requirements.txt
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.3.3
bcrypt==3.1.7
beautifulsoup4==4.9.0
Brlapi==0.6.6
bson==0.5.8
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==7.1.1
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dataclasses==0.7
defer==1.0.6
distro-info==0.18
Django==3.0.6
django-crispy-forms==1.9.0
django-nvd3==0.9.7
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djongo==1.3.2
fake-useragent==0.1.11
Flask==1.1.2
httplib2==0.9.2
idna==2.6
inflection==0.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mtranslate==1.6
mysqlclient==1.4.6
netifaces==0.10.4
numpy==1.18.2
oauth==1.0.1
olefile==0.45.1
opencv-python==4.2.0.32
pandas==1.0.3
paramiko==2.7.1
pexpect==4.2.1
Pillow==7.0.0
prettytable==0.7.2
protobuf==3.0.0
psycopg2-binary==2.8.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
pymongo==3.10.1
pymssql==2.1.4
PyNaCl==1.3.0
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.1
python-debian==0.1.32
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.25.8
usb-creator==0.3.3
useragent==0.1.1
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
